pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Mikhail Orlov <mikhailorlov398@gmail.com>"]
readme = "README.md"
packages = [
  { include = "task_manager" },
]
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Framework :: Django :: 4.1",
    "Framework :: Flake8",
    "Topic :: Education",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]
repository = "https://github.com/Un1que11/python-project-52"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
Django = "^4.1.3"
django-bootstrap4 = "^22.2"
django-filter = "^22.1"
dj-database-url = "0.5.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.2.0"
rollbar = "^0.16.3"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
